+++ src/main/java/com/besson/arknights/ArknightsFurniture.java
@@ -5,6 +5,7 @@ import com.besson.arknights.block.ModBlocks;
 import com.besson.arknights.item.ModItemGroups;
 import com.besson.arknights.item.ModItems;
 import com.besson.arknights.painting.ModPaintingVariants;
+import com.besson.arknights.tag.ModBlockTags;
 import com.besson.arknights.tag.ModItemTags;
 import net.fabricmc.api.ModInitializer;
 
@@ -30,6 +31,7 @@ public class ArknightsFurniture implements ModInitializer {
 		ModPaintingVariants.registerPaintings();
 		ModBlockEntities.registerModBlockEntities();
 		ModItemTags.registerModBlockTags();
+		ModBlockTags.registerModBlockTags();
 
 		LOGGER.info("Hello Fabric world!");
 	}