+++ src/main/java/com/besson/arknights/block/FurnitureProperties.java
@@ -1,0 +1,7 @@
+package com.besson.arknights.block;
+
+import net.minecraft.state.property.BooleanProperty;
+
+public class FurnitureProperties {
+    public static final BooleanProperty TAPED = BooleanProperty.of("taped");
+}