+++ src/main/java/com/besson/arknights/datagen/ModBlockTagsProvider.java
@@ -1,5 +1,7 @@
 package com.besson.arknights.datagen;
 
+import com.besson.arknights.block.ModBlocks;
+import com.besson.arknights.tag.ModBlockTags;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagProvider;
 import net.minecraft.registry.RegistryWrapper;
@@ -13,6 +15,7 @@ public class ModBlockTagsProvider extends FabricTagProvider.BlockTagProvider {
 
     @Override
     protected void configure(RegistryWrapper.WrapperLookup arg) {
-
+        getOrCreateTagBuilder(ModBlockTags.CARTON)
+                .add(ModBlocks.CARTON);
     }
 }